requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
six==1.15.0
huey==2.3.0
redis==3.5.3
django-crispy-forms==1.11.1
Pillow==8.1.2
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0